feat: Introduce getObstacleFromElement function for converting circuit elements into obstacles and refactor related imports
